qcacld-2.0: Convert beacon data rate to hw code
User configures the beacon data rate in multiples of 100 Kbps
and FW expects the data rate in the form of hw rate codes.
So convert the data rates to hw rate code before configuring it to
firmware.

CRs-Fixed: 1020078

+++ CORE/SAP/inc/sapApi.h
@@ -544,7 +544,7 @@ typedef struct sap_Config {
     uint8_t   sap_dot11mc;      /* Specify if 11MC is enabled or disabled*/
     enum sap_acs_dfs_mode acs_dfs_mode;
 
-    uint8_t beacon_tx_rate;
+    uint16_t beacon_tx_rate;
     uint8_t *vendor_ie;
     enum vendor_ie_access_policy vendor_ie_access_policy;
     uint16_t sta_inactivity_timeout;
